# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Fresh Portal — контракт инженерного пилота R1
  version: 0.1.0-proposed
  description: |
    НА СОГЛАСОВАНИЕ. Contract only, not a deployed API or production readiness claim.
    Narrow R1 slice of TZ v2.10 §§13.1A,45 and security rules. Existing personal
    login/password + server sessions; no new OIDC. Synthetic A/B only.
    17 operations; business mutations require Idempotency-Key for at least 30 days.
    login/logout are the only mutating exceptions. Every action POST explicitly
    carries expected_entity_version; field PATCH uses field CAS instead.
    Check current authorization before schema/CAS and before idempotent replay.
    Hidden object is 404; forbidden explicit branch filter/action is 403.
    Commit mutation + audit + outbox + successful idempotency result atomically.
    All responses include X-Request-ID and Cache-Control:no-store.
    Error envelope code,message,request_id,details also applies to malformed JSON
    (422), gateway errors and quota errors. All timestamps returned in UTC with Z.
    423 is reserved for the general restriction contract; sanctions disabled in R1.
servers:
  - url: /api/v1
...
security:
  - sessionCookie: []
paths:
...
  /work-items:
    get:
      tags: [WorkItems]
      operationId: listWorkItems
      summary: Список только доступных задач
      description: |
        RM sees granted branch items; RF sees only own assigned items in granted
        branch. No implicit DIVISION/NETWORK expansion. Forbidden explicit org
        filter returns 403, not rows/count from another branch. No total count.
        Keyset ascending (created_at,id); cursor bound to actor/filter/permissions.
      x-requirement-ids: [R1-02]
      x-permission: work_item.read
      x-events: []
      parameters:
        - $ref: '#/components/parameters/OrgFilter'
        - name: status
          in: query
          schema: {$ref: '#/components/schemas/WorkItemStatus'}
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Scope-filtered page (empty page is valid).
          headers:
            X-Request-ID: {$ref: '#/components/headers/RequestId'}
            Cache-Control: {$ref: '#/components/headers/NoStore'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/WorkItemPage'}
              example: {items: [], next_cursor: null}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '423': {$ref: '#/components/responses/Locked'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '503': {$ref: '#/components/responses/Unavailable'}
    post:
      tags: [WorkItems]
      operationId: createWorkItem
      summary: Создать DRAFT из pilot_task_v1
      description: |
        RM grant in requested org. Pin sole template version. Title/due_at/org
        immutable in R1. Initial field null, field_version=1, entity_version=1.
        due_at is required UTC Z; past deadline allowed. Server assigns actor/IDs.
      x-requirement-ids: [R1-02, R1-03, R1-11, R1-12]
      x-permission: work_item.create
      x-events: [work_item.created]
      security: [{sessionCookie: [], csrfToken: []}]
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateWorkItemRequest'}
            example:
              org_unit_id: '00000000-0000-4000-8000-00000000000a'
              template_code: pilot_task_v1
              title: 'Проверить тестовый сценарий'
              due_at: '2026-09-16T12:00:00Z'
      responses:
        '201':
          description: Created and committed. Same key+payload replays same 201 body.
          headers:
            X-Request-ID: {$ref: '#/components/headers/RequestId'}
            Cache-Control: {$ref: '#/components/headers/NoStore'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/WorkItem'}
              examples:
                draft: {$ref: '#/components/examples/DraftWorkItem'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '423': {$ref: '#/components/responses/Locked'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /work-items/{work_item_id}:
    parameters:
      - $ref: '#/components/parameters/WorkItemId'
    get:
      tags: [WorkItems]
      operationId: getWorkItem
      summary: Карточка с текущими entity/field/submission versions
      description: RM scope or RF-own; unknown or hidden object returns indistinguishable 404.
      x-requirement-ids: [R1-02, R1-05]
      x-permission: work_item.read
      x-events: []
      responses:
        '200': {$ref: '#/components/responses/WorkItemOk'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '423': {$ref: '#/components/responses/Locked'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /work-items/{work_item_id}/assign:
    parameters:
      - $ref: '#/components/parameters/WorkItemId'
    post:
      tags: [WorkItems]
      operationId: assignWorkItem
      summary: DRAFT → ASSIGNED
      description: |
        RM same org; target active RF with current grant in the same org.
        No reassignment. Invalid/hidden/inactive target uniformly returns 422
        ASSIGNEE_INELIGIBLE without identifying another branch's user.
      x-requirement-ids: [R1-02, R1-04, R1-11, R1-12]
      x-permission: work_item.assign
      x-events: [work_item.assigned]
      security: [{sessionCookie: [], csrfToken: []}]
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AssignRequest'}
            example: {expected_entity_version: 1, assignee_user_id: '10000000-0000-4000-8000-000000000002'}
      responses:
        '200': {$ref: '#/components/responses/WorkItemOk'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '423': {$ref: '#/components/responses/Locked'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /work-items/{work_item_id}/start:
    parameters:
      - $ref: '#/components/parameters/WorkItemId'
    post:
      tags: [WorkItems]
      operationId: startWorkItem
      summary: ASSIGNED → IN_PROGRESS
      description: Only assigned active RF with current grant in this branch; no acting users in R1.
      x-requirement-ids: [R1-04, R1-11, R1-12]
      x-permission: work_item.start
      x-events: [work_item.started]
      security: [{sessionCookie: [], csrfToken: []}]
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/VersionRequest'}
            example: {expected_entity_version: 2}
      responses:
        '200': {$ref: '#/components/responses/WorkItemOk'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '423': {$ref: '#/components/responses/Locked'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /work-items/{work_item_id}/fields:
    parameters:
      - $ref: '#/components/parameters/WorkItemId'
    patch:
      tags: [WorkItems]
      operationId: patchWorkItemFields
      summary: CAS сохранённого completion_summary
      description: |
        RF-own, ASSIGNED/IN_PROGRESS only. Exactly one change, expected field
        version, not entity CAS. Serialize with transitions; success increments
        field_version and entity_version once, even if same text with a new key.
        Whitespace-only/null forbidden, max4000 Unicode code points proposed.
        Return full committed card; 409 rolls back whole patch. Cannot change org,
        title, due_at, blocked flag, actor, status, template or other field.
      x-requirement-ids: [R1-05, R1-11, R1-12]
      x-permission: work_item.fields.write
      x-events: [work_item.fields_patched]
      security: [{sessionCookie: [], csrfToken: []}]
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/PatchFieldsRequest'}
            example:
              changes:
                - {field_path: completion_summary, expected_version: 1, new_value: 'Тестовый сценарий проверен.'}
      responses:
        '200': {$ref: '#/components/responses/WorkItemOk'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '423': {$ref: '#/components/responses/Locked'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /work-items/{work_item_id}/submit:
    parameters:
      - $ref: '#/components/parameters/WorkItemId'
    post:
      tags: [WorkItems]
      operationId: submitWorkItem
      summary: ASSIGNED/IN_PROGRESS → SUBMITTED с immutable snapshot
      description: |
        RF-own; !is_blocked; persisted completion_summary nonblank. Server clock
        submitted_at; next revision and fresh submission_id generated atomically.
        Requires acceptance always true for pilot_task_v1. Past due allowed with
        LATE marker; no client timestamp or unsaved field accepted in this body.
        UI must first await PATCH commit and use returned entity_version.
      x-requirement-ids: [R1-06, R1-11, R1-12]
      x-permission: work_item.submit
      x-events: [work_item.submitted]
      security: [{sessionCookie: [], csrfToken: []}]
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/VersionRequest'}
            example: {expected_entity_version: 4}
      responses:
        '200': {$ref: '#/components/responses/WorkItemOk'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '423': {$ref: '#/components/responses/Locked'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /work-items/{work_item_id}/accept:
    parameters:
      - $ref: '#/components/parameters/WorkItemId'
    post:
      tags: [WorkItems]
      operationId: acceptWorkItem
      summary: SUBMITTED → COMPLETED
      description: |
        RM current branch grant; reviewer must differ from both assignee and
        snapshot submitted_by, including dual-role users. Exact current submission
        id/revision AND entity version required. Preserve snapshot unchanged.
      x-requirement-ids: [R1-02, R1-07, R1-11, R1-12]
      x-permission: work_item.accept
      x-events: [work_item.accepted]
      security: [{sessionCookie: [], csrfToken: []}]
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ReviewRequest'}
            example: {expected_entity_version: 5, submission_id: '40000000-0000-4000-8000-000000000001', submission_revision: 1}
      responses:
        '200': {$ref: '#/components/responses/WorkItemOk'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '423': {$ref: '#/components/responses/Locked'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /work-items/{work_item_id}/rework:
    parameters:
      - $ref: '#/components/parameters/WorkItemId'
    post:
      tags: [WorkItems]
      operationId: reworkWorkItem
      summary: SUBMITTED → IN_PROGRESS с причиной
      description: |
        Same independent reviewer guards as accept. Exact submission id/revision
        and entity version. Reason nonblank max500 proposed; rework_count+1.
        Preserve original due_at and immutable submission, next submit is revision+1.
      x-requirement-ids: [R1-07, R1-11, R1-12]
      x-permission: work_item.rework
      x-events: [work_item.rework_requested]
      security: [{sessionCookie: [], csrfToken: []}]
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ReworkRequest'}
            example: {expected_entity_version: 5, submission_id: '40000000-0000-4000-8000-000000000001', submission_revision: 1, reason: 'Уточнить результат проверки.'}
      responses:
        '200': {$ref: '#/components/responses/WorkItemOk'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '423': {$ref: '#/components/responses/Locked'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /work-items/{work_item_id}/cancel:
    parameters:
      - $ref: '#/components/parameters/WorkItemId'
    post:
      tags: [WorkItems]
      operationId: cancelWorkItem
      summary: DRAFT/ASSIGNED/IN_PROGRESS/SUBMITTED → CANCELLED
      description: |
        RM branch grant, reason mandatory. No parent/children in pilot, thus no
        cascade preview needed only in R1. Clear blocked flag/reason, audit before
        preserves obstacle. Keep submissions/assignee/deadline; CANCELLED terminal.
      x-requirement-ids: [R1-08, R1-11, R1-12]
      x-permission: work_item.cancel
      x-events: [work_item.cancelled]
      security: [{sessionCookie: [], csrfToken: []}]
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ReasonRequest'}
            example: {expected_entity_version: 3, reason: 'Тестовый сценарий отменён.'}
      responses:
        '200': {$ref: '#/components/responses/WorkItemOk'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '423': {$ref: '#/components/responses/Locked'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /work-items/{work_item_id}/reopen:
    parameters:
      - $ref: '#/components/parameters/WorkItemId'
    post:
      tags: [WorkItems]
      operationId: reopenWorkItem
      summary: COMPLETED → IN_PROGRESS с причиной
      description: |
        RM branch grant; previous assignee still an active RF with current branch
        grant, else 422 ASSIGNEE_INELIGIBLE. No reassignment. Keep accepted snapshot,
        original deadline and rework_count; subsequent submit creates new revision.
      x-requirement-ids: [R1-08, R1-11, R1-12]
      x-permission: work_item.reopen
      x-events: [work_item.reopened]
      security: [{sessionCookie: [], csrfToken: []}]
      parameters:
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ReasonRequest'}
            example: {expected_entity_version: 6, reason: 'Повторная проверка.'}
      responses:
        '200': {$ref: '#/components/responses/WorkItemOk'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '423': {$ref: '#/components/responses/Locked'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '503': {$ref: '#/components/responses/Unavailable'}
  /work-items/{work_item_id}/history:
    parameters:
      - $ref: '#/components/parameters/WorkItemId'
    get:
      tags: [WorkItems]
      operationId: getWorkItemHistory
      summary: Версионированная история и snapshot каждой сдачи
      description: |
        RM scope or RF-own at read time. Ordered ascending (aggregate_version,event_id).
        Only APPLIED business events for this work item, not rejected attempts,
        auth/security audit, IP, user-agent or hidden fields. Submission snapshot
        appears on SUBMIT entry; accept/rework reference reviewed id/revision.
      x-requirement-ids: [R1-02, R1-09]
      x-permission: work_item.history.read
      x-events: []
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Authorized append-only history page.
          headers:
            X-Request-ID: {$ref: '#/components/headers/RequestId'}
            Cache-Control: {$ref: '#/components/headers/NoStore'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/HistoryPage'}
              example:
                items:
                  - event_id: '50000000-0000-4000-8000-000000000001'
                    event_type: work_item.created
                    aggregate_version: 1
                    occurred_at: '2026-09-15T07:41:00Z'
                    actor_id: '10000000-0000-4000-8000-000000000001'
                    from_status: null
                    to_status: DRAFT
                    reason: null
                    field_change: null
                    submission: null
                    reviewed_submission_id: null
                    reviewed_submission_revision: null
                next_cursor: null
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '423': {$ref: '#/components/responses/Locked'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '503': {$ref: '#/components/responses/Unavailable'}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-fresh_session
      description: Existing portal server session, opaque Secure HttpOnly cookie; never a role-bearing JWT.
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Session-bound token from login/getMe; required together with cookie on all non-login unsafe operations.
  headers:
    RequestId:
      schema: {type: string, format: uuid}
      description: Server trace UUID; equals error body's request_id if error.
    NoStore:
      schema: {type: string, enum: [no-store]}
      description: Do not cache personal content, tokens or errors.
...
  parameters:
    Origin:
      name: Origin
      in: header
      required: true
      description: Exact approved HTTPS origin; allowlist hostname TBD before R1 starts. Missing/null/other origin gives 403.
      schema: {type: string, format: uri}
      example: https://pilot.example.invalid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Unique(actor,operation,key); ≥30d. Fingerprint includes target/path and
        canonical JSON payload. Same payload returns saved status/body only after
        fresh read authorization. Changed payload/target ->409. Missing key ->422.
        All business mutations, not login/logout. Do not retry expired create blindly.
      schema: {type: string, minLength: 16, maxLength: 128, pattern: '^[A-Za-z0-9._:-]+$'}
      example: 'pilot-r1-20260915-000001'
    WorkItemId:
      name: work_item_id
      in: path
      required: true
      schema: {$ref: '#/components/schemas/Uuid'}
    OrgFilter:
      name: org_unit_id
      in: query
      required: false
      description: Optional explicit authorized branch. Omitted means only effective scope, never network access.
      schema: {$ref: '#/components/schemas/Uuid'}
    Limit:
      name: limit
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 50}
    Cursor:
      name: cursor
      in: query
      description: Opaque actor/filter/permission-bound keyset cursor, proposed TTL15min; not an access capability.
      schema: {type: string, minLength: 1, maxLength: 512}
  schemas:
    Uuid:
      type: string
      format: uuid
...
    WorkItemStatus:
      type: string
      enum: [DRAFT, ASSIGNED, IN_PROGRESS, SUBMITTED, COMPLETED, CANCELLED]
...
    CreateWorkItemRequest:
      type: object
      additionalProperties: false
      required: [org_unit_id, template_code, title, due_at]
      properties:
        org_unit_id: {$ref: '#/components/schemas/Uuid'}
        template_code: {type: string, enum: [pilot_task_v1]}
        title: {type: string, minLength: 1, maxLength: 200, pattern: '\S'}
        due_at: {$ref: '#/components/schemas/UtcTimestamp'}
...
    WorkItemPage:
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items: {type: array, maxItems: 100, items: {$ref: '#/components/schemas/WorkItem'}}
        next_cursor: {$ref: '#/components/schemas/NextCursor'}
...
  responses:
...
    Unauthorized:
      description: 401 absent/expired/revoked session; login uses uniform INVALID_CREDENTIALS.
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestId'}
        Cache-Control: {$ref: '#/components/headers/NoStore'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {code: UNAUTHENTICATED, message: 'Требуется персональный вход.', request_id: '70000000-0000-4000-8000-000000000001', details: {}}
    Forbidden:
      description: 403 permission/field/CSRF/Origin/self-review denied; no forbidden data in details.
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestId'}
        Cache-Control: {$ref: '#/components/headers/NoStore'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {code: FORBIDDEN, message: 'Действие не разрешено.', request_id: '70000000-0000-4000-8000-000000000002', details: {}}
    NotFound:
      description: 404 missing or hidden object/notification, indistinguishable response.
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestId'}
        Cache-Control: {$ref: '#/components/headers/NoStore'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {code: NOT_FOUND, message: 'Объект не найден.', request_id: '70000000-0000-4000-8000-000000000003', details: {}}
    Conflict:
      description: 409 CAS/entity/submission/key/schema conflict; whole request rolled back. Values only after field authorization.
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestId'}
        Cache-Control: {$ref: '#/components/headers/NoStore'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          examples:
            field:
              value:
                code: FIELD_VERSION_CONFLICT
                message: 'Поле изменилось, требуется явное разрешение конфликта.'
                request_id: '70000000-0000-4000-8000-000000000004'
                details:
                  current_entity_version: 4
                  conflicts: [{field_path: completion_summary, expected_version: 1, current_version: 2, current_value: 'Сохранённый результат.'}]
            submission:
              value:
                code: SUBMISSION_CONFLICT
                message: 'Версия сдачи устарела.'
                request_id: '70000000-0000-4000-8000-000000000005'
                details: {current_submission_id: '40000000-0000-4000-8000-000000000002', current_submission_revision: 2}
    Unprocessable:
      description: 422 malformed JSON/schema/business guard/invalid cursor; no partial effects.
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestId'}
        Cache-Control: {$ref: '#/components/headers/NoStore'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          examples:
            transition:
              value: {code: INVALID_TRANSITION, message: 'Переход из текущего состояния запрещён.', request_id: '70000000-0000-4000-8000-000000000006', details: {current_status: CANCELLED}}
            blocked:
              value: {code: WORK_ITEM_BLOCKED, message: 'Нельзя сдать заблокированную задачу.', request_id: '70000000-0000-4000-8000-000000000007', details: {}}
    Locked:
      description: 423 reserved restriction contract; no sanction activation or legal restriction workflow in R1.
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestId'}
        Cache-Control: {$ref: '#/components/headers/NoStore'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {code: RESTRICTION_ACTIVE, message: 'Действует ограничение доступа.', request_id: '70000000-0000-4000-8000-000000000008', details: {}}
    TooManyRequests:
      description: 429 server quota exceeded; Retry-After required, retry with same idempotency key.
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestId'}
        Cache-Control: {$ref: '#/components/headers/NoStore'}
        Retry-After: {$ref: '#/components/headers/RetryAfter'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {code: RATE_LIMITED, message: 'Повторите позже.', request_id: '70000000-0000-4000-8000-000000000009', details: {retry_after_seconds: 30}}
    Unavailable:
      description: 503 transient/unconfirmed outcome; Retry-After required, repeat same key to resolve committed result.
      headers:
        X-Request-ID: {$ref: '#/components/headers/RequestId'}
        Cache-Control: {$ref: '#/components/headers/NoStore'}
        Retry-After: {$ref: '#/components/headers/RetryAfter'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
          example: {code: TEMPORARILY_UNAVAILABLE, message: 'Сервис временно недоступен.', request_id: '70000000-0000-4000-8000-000000000010', details: {retry_after_seconds: 30}}
...